Fixing models that had BIGINT to go to Integer. Also fixed DateTime in Users.

+++ bot/database/models/profile.py
@@ -5,4 +5,4 @@ class Profile(db.Model):
     __tablename__ = "Profile"
 
     user_id = db.Column(db.BIGINT, primary_key=True)
-    badge_id = db.Column(db.BIGINT)
+    badge_id = db.Column(db.Integer)
